requirements.txt
openai>=1.0.0
arize-phoenix>=8.8.0
arize-phoenix-otel>=0.8.0
openinference-instrumentation-langchain
python-dotenv>=1.0.0
duckdb>=0.9.0
openinference-instrumentation>=0.1.21
langgraph>=0.0.10
langchain>=0.1.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0
tqdm>=4.65.0
pydantic>=2.0.0 
langchain_openai>=0.3.0